fix: remove Babel, use combined CJS/ESM approach
* approach 1 from https://nodejs.org/api/esm.html#esm_writing_dual_packages_while_avoiding_or_minimizing_hazards
