+++ packages/contrast-colors/test/createScale.test.js
@@ -9,7 +9,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { createScale } from '../index.js';
+const { createScale } = require('../index.cjs');
 
 test('should generate 8 colors in Lab', function() {
   let scale = createScale({swatches: 8, colorKeys: ['#CCFFA9', '#FEFEC5', '#5F0198'], colorspace: 'LAB', shift: 1, fullScale: true});
