+++ packages/contrast-colors/test/generateBaseScale.test.js
@@ -9,7 +9,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { generateBaseScale } from '../index.js';
+const { generateBaseScale } = require('../index.cjs');
 
 // Test simple generation in all color spaces
 test('should generate 101 grayscale colors in default LAB colorspace)', function() {
