+++ packages/contrast-colors/test/minPositive.test.js
@@ -9,7 +9,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { minPositive } from '../index.js';
+const { minPositive } = require('../index.cjs');
 
 test('should return 1', function() {
   let result = minPositive([1, 1.2, 1.4, 2, 3, 4.5, 6, 8, 12, 21]);;
