+++ packages/contrast-colors/wrapper.mjs
@@ -1,0 +1,13 @@
+import contrastColors from './index.cjs';
+
+export const createScale = contrastColors.createScale;
+export const luminance = contrastColors.luminance;
+export const contrast = contrastColors.contrast;
+export const binarySearch = contrastColors.binarySearch;
+export const generateBaseScale = contrastColors.generateBaseScale;
+export const generateContrastColors = contrastColors.generateContrastColors;
+export const minPositive = contrastColors.minPositive;
+export const ratioName = contrastColors.ratioName;
+export const generateAdaptiveTheme = contrastColors.generateAdaptiveTheme;
+
+export default contrastColors;
